backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
celery==5.4.0
redis==5.1.1
numpy==2.1.2
fpdf2==2.8.1
jinja2==3.1.4
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
